pyproject.toml
[tool.poetry]
name = "image-magician"
version = "0.1.0"
description = ""
authors = ["Luca Zappaterra <luca.zappaterra@me.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
streamlit = "^1.41.1"
python-dotenv = "^1.0.1"
pillow = "^11.0.0"
diffusers = "^0.31.0"
transformers = "^4.47.1"
opencv-python = "^4.10.0.84"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
